Initial rust CI setup

+++ src/sources/memory.rs
@@ -21,7 +21,7 @@ impl MemorySource {
 
     pub async fn read(&mut self, offset: u64, buf: &mut [u8]) -> Result<usize, Error> {
         let end = std::cmp::min(self.buffer.len(), offset as usize + buf.len());
-        buf[..(end - offset as usize)].copy_from_slice(&self.buffer[offset as usize..end as usize]);
+        buf[..(end - offset as usize)].copy_from_slice(&self.buffer[offset as usize..end]);
         self.stats.read_counts += 1;
         Ok(end - offset as usize)
     }
